requirements.txt
streamlit
pandas
matplotlib
fpdf2
sqlite-utils
cryptography
python-dateutil
openpyxl

